UDT-23 refactor: Project Entity Package Structure Change
