+++ src/main/java/com/example/udtbe/domain/admin/repository/AdminRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.admin.repository;
 
-import com.example.udtbe.entity.Member;
+import com.example.udtbe.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminRepository extends JpaRepository<Member, Long> {
