+++ src/main/java/com/example/udtbe/domain/content/entity/Cast.java
@@ -1,4 +1,4 @@
-package com.example.udtbe.entity;
+package com.example.udtbe.domain.content.entity;
 
 import static lombok.AccessLevel.PRIVATE;
 import static lombok.AccessLevel.PROTECTED;
@@ -15,31 +15,36 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "country")
+@Table(name = "cast")
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class Country extends TimeBaseEntity {
+public class Cast extends TimeBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "country_id")
+    @Column(name = "cast_id")
     private Long id;
 
-    @Column(name = "country_name", nullable = false)
-    private String countryName;
+    @Column(name = "cast_name", nullable = false)
+    private String castName;
+
+    @Column(name = "cast_image_url")
+    private String castImageUrl;
 
     @Column(name = "is_deleted", nullable = false)
     private boolean isDeleted;
 
     @Builder(access = PRIVATE)
-    private Country(String countryName, boolean isDeleted) {
-        this.countryName = countryName;
+    private Cast(String castName, String castImageUrl, boolean isDeleted) {
+        this.castName = castName;
+        this.castImageUrl = castImageUrl;
         this.isDeleted = isDeleted;
     }
 
-    public static Country of(String countryName, boolean isDeleted) {
-        return Country.builder()
-                .countryName(countryName)
+    public static Cast of(String castName, String castImageUrl, boolean isDeleted) {
+        return Cast.builder()
+                .castName(castName)
+                .castImageUrl(castImageUrl)
                 .isDeleted(isDeleted)
                 .build();
     }
