+++ src/main/java/com/example/udtbe/domain/content/entity/Genre.java
@@ -1,4 +1,4 @@
-package com.example.udtbe.entity;
+package com.example.udtbe.domain.content.entity;
 
 import static jakarta.persistence.ConstraintMode.NO_CONSTRAINT;
 import static jakarta.persistence.EnumType.STRING;
@@ -6,7 +6,7 @@ import static jakarta.persistence.FetchType.LAZY;
 import static lombok.AccessLevel.PRIVATE;
 import static lombok.AccessLevel.PROTECTED;
 
-import com.example.udtbe.entity.enums.FeedbackType;
+import com.example.udtbe.domain.content.entity.enums.GenreType;
 import com.example.udtbe.global.entity.TimeBaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -23,50 +23,41 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "feedback")
+@Table(name = "genre")
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class Feedback extends TimeBaseEntity {
+public class Genre extends TimeBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "feedback_id")
+    @Column(name = "genre_id")
     private Long id;
 
     @Enumerated(value = STRING)
-    @Column(name = "feedback_type", nullable = false)
-    private FeedbackType feedbackType;
+    @Column(name = "genre_type", nullable = false)
+    private GenreType genreType;
 
     @Column(name = "is_deleted", nullable = false)
     private boolean isDeleted;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "member_id",
+    @JoinColumn(name = "category_id",
             nullable = false,
             foreignKey = @ForeignKey(NO_CONSTRAINT))
-    private Member member;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "content_id",
-            nullable = false,
-            foreignKey = @ForeignKey(NO_CONSTRAINT))
-    private Content content;
+    private Category category;
 
     @Builder(access = PRIVATE)
-    private Feedback(FeedbackType feedbackType, boolean isDeleted, Member member, Content content) {
-        this.feedbackType = feedbackType;
+    private Genre(GenreType genreType, boolean isDeleted, Category category) {
+        this.genreType = genreType;
         this.isDeleted = isDeleted;
-        this.member = member;
-        this.content = content;
+        this.category = category;
     }
 
-    public static Feedback of(FeedbackType feedbackType, boolean isDeleted, Member member,
-            Content content) {
-        return Feedback.builder()
-                .feedbackType(feedbackType)
+    public static Genre of(GenreType genreType, boolean isDeleted, Category category) {
+        return Genre.builder()
+                .genreType(genreType)
                 .isDeleted(isDeleted)
-                .member(member)
-                .content(content)
+                .category(category)
                 .build();
     }
 }
