+++ src/main/java/com/example/udtbe/domain/content/entity/Platform.java
@@ -1,4 +1,4 @@
-package com.example.udtbe.entity;
+package com.example.udtbe.domain.content.entity;
 
 import static lombok.AccessLevel.PRIVATE;
 import static lombok.AccessLevel.PROTECTED;
@@ -15,36 +15,31 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "cast")
+@Table(name = "platform")
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class Cast extends TimeBaseEntity {
+public class Platform extends TimeBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "cast_id")
+    @Column(name = "platform_id")
     private Long id;
 
-    @Column(name = "cast_name", nullable = false)
-    private String castName;
-
-    @Column(name = "cast_image_url")
-    private String castImageUrl;
+    @Column(name = "platform_name", nullable = false)
+    private String platformName;
 
     @Column(name = "is_deleted", nullable = false)
     private boolean isDeleted;
 
     @Builder(access = PRIVATE)
-    private Cast(String castName, String castImageUrl, boolean isDeleted) {
-        this.castName = castName;
-        this.castImageUrl = castImageUrl;
+    private Platform(String platformName, boolean isDeleted) {
+        this.platformName = platformName;
         this.isDeleted = isDeleted;
     }
 
-    public static Cast of(String castName, String castImageUrl, boolean isDeleted) {
-        return Cast.builder()
-                .castName(castName)
-                .castImageUrl(castImageUrl)
+    public static Platform of(String platformName, boolean isDeleted) {
+        return Platform.builder()
+                .platformName(platformName)
                 .isDeleted(isDeleted)
                 .build();
     }
