+++ src/main/java/com/example/udtbe/domain/content/entity/enums/FeedbackType.java
@@ -1,4 +1,4 @@
-package com.example.udtbe.entity.enums;
+package com.example.udtbe.domain.content.entity.enums;
 
 import com.example.udtbe.global.exception.RestApiException;
 import com.example.udtbe.global.exception.code.EnumErrorCode;
@@ -8,20 +8,18 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CategoryType {
+public enum FeedbackType {
 
-    MOVIE("영화"),
-    DRAMA("드라마"),
-    ANIMATION("애니메이션"),
-    VARIETY("예능"),
+    LIKE("좋아요"),
+    DISLIKE("싫어요"),
     ;
 
     private final String type;
 
-    public static CategoryType from(String value) {
+    public static FeedbackType from(String value) {
         return Arrays.stream(values())
                 .filter(r -> r.getType().equals(value))
                 .findFirst()
-                .orElseThrow(() -> new RestApiException(EnumErrorCode.CATEGORY_TYPE_NOT_FOUND));
+                .orElseThrow(() -> new RestApiException(EnumErrorCode.FEEDBACK_TYPE_NOT_FOUND));
     }
 }
