+++ src/main/java/com/example/udtbe/domain/content/repository/CastRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.Cast;
+import com.example.udtbe.domain.content.entity.Cast;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CastRepository extends JpaRepository<Cast, Long> {
