+++ src/main/java/com/example/udtbe/domain/content/repository/ContentCastRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.ContentCast;
+import com.example.udtbe.domain.content.entity.ContentCast;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentCastRepository extends JpaRepository<ContentCast, Long> {
