+++ src/main/java/com/example/udtbe/domain/content/repository/ContentCountryRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.ContentCountry;
+import com.example.udtbe.domain.content.entity.ContentCountry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentCountryRepository extends JpaRepository<ContentCountry, Long> {
