+++ src/main/java/com/example/udtbe/domain/content/repository/ContentDirectorRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.ContentDirector;
+import com.example.udtbe.domain.content.entity.ContentDirector;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentDirectorRepository extends JpaRepository<ContentDirector, Long> {
