+++ src/main/java/com/example/udtbe/domain/content/repository/ContentMetadataRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.ContentMetadata;
+import com.example.udtbe.domain.content.entity.ContentMetadata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentMetadataRepository extends JpaRepository<ContentMetadata, Long> {
