+++ src/main/java/com/example/udtbe/domain/content/repository/ContentPlatformRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.ContentPlatform;
+import com.example.udtbe.domain.content.entity.ContentPlatform;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentPlatformRepository extends JpaRepository<ContentPlatform, Long> {
