+++ src/main/java/com/example/udtbe/domain/content/repository/ContentRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.Content;
+import com.example.udtbe.domain.content.entity.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentRepository extends JpaRepository<Content, Long> {
