+++ src/main/java/com/example/udtbe/domain/content/repository/CountryRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.Country;
+import com.example.udtbe.domain.content.entity.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CountryRepository extends JpaRepository<Country, Long> {
