+++ src/main/java/com/example/udtbe/domain/content/repository/DirectorRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.Director;
+import com.example.udtbe.domain.content.entity.Director;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DirectorRepository extends JpaRepository<Director, Long> {
