+++ src/main/java/com/example/udtbe/domain/content/repository/PlatformRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.content.repository;
 
-import com.example.udtbe.entity.Platform;
+import com.example.udtbe.domain.content.entity.Platform;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlatformRepository extends JpaRepository<Platform, Long> {
