+++ src/main/java/com/example/udtbe/domain/survey/repository/SurveyRepository.java
@@ -1,6 +1,6 @@
 package com.example.udtbe.domain.survey.repository;
 
-import com.example.udtbe.entity.Survey;
+import com.example.udtbe.domain.survey.entity.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyRepository extends JpaRepository<Survey, Long> {
