+++ src/test/java/com/example/udtbe/common/fixture/MemberFixture.java
@@ -1,10 +1,10 @@
 package com.example.udtbe.common.fixture;
 
-import static com.example.udtbe.entity.enums.Gender.MAN;
+import static com.example.udtbe.domain.member.entity.enums.Gender.MAN;
 import static lombok.AccessLevel.PRIVATE;
 
-import com.example.udtbe.entity.Member;
-import com.example.udtbe.entity.enums.Role;
+import com.example.udtbe.domain.member.entity.Member;
+import com.example.udtbe.domain.member.entity.enums.Role;
 import java.time.LocalDateTime;
 import lombok.NoArgsConstructor;
